+++ ww/base/widle_table.c
@@ -39,32 +39,32 @@ void idleCallBack(wtimer_t *timer);
 widle_table_t *idleTableCreate(wloop_t *loop)
 {
     // assert(sizeof(widle_table_t) <= kCpuLineCacheSize); promotion to 128 bytes
-    int64_t memsize = (int64_t) sizeof(widle_table_t);
+    size_t memsize = sizeof(widle_table_t);
     // ensure we have enough space to offset the allocation by line cache (for alignment)
     MUSTALIGN2(memsize + ((kCpuLineCacheSize + 1) / 2), kCpuLineCacheSize);
-    memsize = (int64_t) ALIGN2(memsize + ((kCpuLineCacheSize + 1) / 2), kCpuLineCacheSize);
+    memsize = ALIGN2(memsize + ((kCpuLineCacheSize + 1) / 2), kCpuLineCacheSize);
 
     // check for overflow
-    if (memsize < (int64_t) sizeof(widle_table_t))
+    if (memsize < sizeof(widle_table_t))
     {
         printError("buffer size out of range");
         exit(1);
     }
 
     // allocate memory, placing widle_table_t at a line cache address boundary
-    uintptr_t ptr = (uintptr_t) memoryAllocate((size_t) memsize);
+    uintptr_t ptr = (uintptr_t) memoryAllocate(memsize);
 
     // align c to line cache boundary
     MUSTALIGN2(ptr, kCpuLineCacheSize);
 
     widle_table_t *newtable = (widle_table_t *) ALIGN2(ptr, kCpuLineCacheSize); // NOLINT
 
-    *newtable = (widle_table_t) {.memptr         = ptr,
-                                 .loop           = loop,
-                                 .idle_handle    = wtimerAdd(loop, idleCallBack, 1000, INFINITE),
-                                 .hqueue         = heapq_idles_t_with_capacity(kVecCap),
-                                 .hmap           = hmap_idles_t_with_capacity(kVecCap),
-                                 .last_update_ms = wloopNowMS(loop)};
+    *newtable = (widle_table_t){.memptr         = ptr,
+                                .loop           = loop,
+                                .idle_handle    = wtimerAdd(loop, idleCallBack, 1000, INFINITE),
+                                .hqueue         = heapq_idles_t_with_capacity(kVecCap),
+                                .hmap           = hmap_idles_t_with_capacity(kVecCap),
+                                .last_update_ms = wloopNowMS(loop)};
 
     mutexInit(&(newtable->mutex));
     weventSetUserData(newtable->idle_handle, newtable);
@@ -77,12 +77,12 @@ idle_item_t *idleItemNew(widle_table_t *self, hash_t key, void *userdata, Expire
     idle_item_t *item = memoryAllocate(sizeof(idle_item_t));
     mutexLock(&(self->mutex));
 
-    *item = (idle_item_t) {.expire_at_ms = wloopNowMS(getWorkerLoop(tid)) + age_ms,
-                           .hash         = key,
-                           .tid          = tid,
-                           .userdata     = userdata,
-                           .cb           = cb,
-                           .table        = self};
+    *item = (idle_item_t){.expire_at_ms = wloopNowMS(getWorkerLoop(tid)) + age_ms,
+                          .hash         = key,
+                          .tid          = tid,
+                          .userdata     = userdata,
+                          .cb           = cb,
+                          .table        = self};
 
     if (! hmap_idles_t_insert(&(self->hmap), item->hash, item).inserted)
     {
