+++ ww/net/tunnel.c
@@ -174,18 +174,18 @@ void tunnelDefaultOnStart(tunnel_t *t)
 }
 
 // Creates a new tunnel instance
-tunnel_t *tunnelCreate(node_t *node, uint16_t tstate_size, uint16_t lstate_size)
+tunnel_t *tunnelCreate(node_t *node, uint32_t tstate_size, uint32_t lstate_size)
 {
     tstate_size = tunnelGetCorrectAllignedStateSize(tstate_size);
     lstate_size = tunnelGetCorrectAllignedLineStateSize(lstate_size);
 
-    int64_t tsize = sizeof(tunnel_t) + tstate_size;
+    size_t tsize = sizeof(tunnel_t) + tstate_size;
     // ensure we have enough space to offset the allocation by line cache (for alignment)
     MUSTALIGN2(tsize + ((kCpuLineCacheSize + 1) / 2), kCpuLineCacheSize);
-    tsize = (int64_t) ALIGN2(tsize + ((kCpuLineCacheSize + 1) / 2), kCpuLineCacheSize);
+    tsize = ALIGN2(tsize + ((kCpuLineCacheSize + 1) / 2), kCpuLineCacheSize);
 
     // allocate memory, placing tunnel_t at a line cache address boundary
-    uintptr_t ptr = (uintptr_t) memoryAllocate((size_t)tsize);
+    uintptr_t ptr = (uintptr_t) memoryAllocate(tsize);
     if (ptr == 0x0)
     {
         // Handle memory allocation failure
